+++ example/app/view/account/Main.js
@@ -2,6 +2,7 @@ Ext.define("StackMobSenchaTouchDemo.view.account.Main", {
     extend: 'Ext.Panel',
 
     requires: [
+        'StackMobSenchaTouchDemo.view.account.CreateUser',
         'StackMobSenchaTouchDemo.view.account.Login',
         'StackMobSenchaTouchDemo.view.account.Profile',
         'StackMobSenchaTouchDemo.view.account.ResetPassword'
@@ -19,6 +20,8 @@ Ext.define("StackMobSenchaTouchDemo.view.account.Main", {
             xclass: 'StackMobSenchaTouchDemo.view.account.Profile'
         },{
             xclass: 'StackMobSenchaTouchDemo.view.account.ResetPassword'
+        },{
+            xclass: 'StackMobSenchaTouchDemo.view.account.CreateUser'
         }]
     }
 });